+++ src/pieces/king.rs
@@ -8,43 +8,43 @@ pub fn generate_pseudo_legal_moves(from: usize, board: &Board, result: &mut Vec<
     let king = board.pieces[from];
     let king_color = get_piece_color(king);
     let mut to = signed_from + 7;
-    if to < 64 && to & 7 < from_file && board.can_be_moved(to as usize, king_color) {
-        result.push(new_move(from, to as usize, board));
+    if to < 64 && to & 7 < from_file && board.can_be_moved(to, king_color) {
+        result.push(new_move(from, to, board));
     }
 
     to = signed_from + 8;
-    if to < 64 && board.can_be_moved(to as usize, king_color) {
-        result.push(new_move(from, to as usize, board));
+    if to < 64 && board.can_be_moved(to, king_color) {
+        result.push(new_move(from, to, board));
     }
 
     to = signed_from + 9;
-    if to < 64 && to & 7 > from_file && board.can_be_moved(to as usize, king_color) {
-        result.push(new_move(from, to as usize, board));
+    if to < 64 && to & 7 > from_file && board.can_be_moved(to, king_color) {
+        result.push(new_move(from, to, board));
     }
 
     to = signed_from + 1;
-    if to & 7 > from_file && board.can_be_moved(to as usize, king_color) {
-        result.push(new_move(from, to as usize, board));
+    if to & 7 > from_file && board.can_be_moved(to, king_color) {
+        result.push(new_move(from, to, board));
     }
 
     to = signed_from - 7;
-    if to >= 0 && to & 7 > from_file && board.can_be_moved(to as usize, king_color) {
-        result.push(new_move(from, to as usize, board));
+    if to >= 0 && to & 7 > from_file && board.can_be_moved(to, king_color) {
+        result.push(new_move(from, to, board));
     }
 
     to = signed_from - 8;
-    if to >= 0 && board.can_be_moved(to as usize, king_color) {
-        result.push(new_move(from, to as usize, board));
+    if to >= 0 && board.can_be_moved(to, king_color) {
+        result.push(new_move(from, to, board));
     }
 
     to = signed_from - 9;
-    if to >= 0 && to & 7 < from_file && board.can_be_moved(to as usize, king_color) {
-        result.push(new_move(from, to as usize, board));
+    if to >= 0 && to & 7 < from_file && board.can_be_moved(to, king_color) {
+        result.push(new_move(from, to, board));
     }
 
     to = signed_from - 1;
-    if to & 7 < from_file && board.can_be_moved(to as usize, king_color) {
-        result.push(new_move(from, to as usize, board));
+    if to & 7 < from_file && board.can_be_moved(to, king_color) {
+        result.push(new_move(from, to, board));
     }
 
     if board.is_castling_queens_side_pseudo_legal(king_color) {
